raspberry_pi/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pi-ina219>=1.4.0
